Cleared out JSON related code, not needed for API

+++ db.py
@@ -1,6 +1,4 @@
 from pymongo import MongoClient
-import json
-import os
 
 # Local/URI connection
 client = MongoClient("mongodb://localhost:27017")
@@ -16,7 +14,6 @@ def add_book(book_data):
     # Validate data --> add book
     if book_data != {}:
         collection.insert_one(book_data)
-        add_book_json(book_data)
 
 
 
@@ -32,7 +29,7 @@ def delete_all_docs():
     collection.delete_many({})
 
     # Check if deletion was successful by retrieving and counting the collection from get_all_books()
-    return "\n All entries deleted \n" if len(get_all_books()) <= 0 else "\n Could not remove any books \n" # Practicing ternary operators
+    return "\n All entries deleted \n" if len(get_all_books()) == 0 else "\n Could not remove any books \n" # Practicing ternary operators
 
 
 
@@ -40,11 +37,12 @@ def delete_all_docs():
 def delete_book(query):
     #Get the title of the book - only for confirmation message to user (instead of getting the ISBN nr)
     title = search_books(query)
+    if not title:
+        return "-- The book could not be found. Check the ISBN number. --"
 
     # Count the amount of books before and after removal
     count_before = get_all_books()
     collection.delete_one({"isbn": query})
-    remove_book_from_json(query)
     count_after = get_all_books()
 
     # Comparison to see if the book is removed
@@ -64,53 +62,6 @@ def search_books(query):
     # Search for all books by the author (case-sensitive)
     else:
         # $regex: f"^{query}$" searches for the exakt match
-        return list(collection.find({"author": {"$regex": f"^{query}$", "$options": "i"}})) #   ^ = beginning of string, $ = end of string
-
-        
-
-''' Function to sync added book between database and json file '''
-def add_book_json(new_book):
-    json_path = os.path.join(os.path.dirname(__file__), "books.json")
-
-    try:
-        # Read existing books
-        with open(json_path, "r", encoding="utf-8") as f:
-            books = json.load(f)
-    except (FileNotFoundError, json.JSONDecodeError):
-        books = []
-
-    # Remove MongoDBs ObjectId - Raises TypeError: Object of type ObjectId is not JSON serializable
-    if "_id" in new_book:
-        del new_book["_id"]
-
-    # Add the new book
-    books.append(new_book)
-
-    # Write to file
-    with open(json_path, "w", encoding="utf-8") as f:
-        json.dump(books, f, ensure_ascii=False, indent=4)
-
-
-
-''' Function to sync deleted book between database and json file '''
-def remove_book_from_json(deleted_book):
-    json_path = os.path.join(os.path.dirname(__file__), "books.json")
-
-    try:
-        # Read existing books
-        with open(json_path, "r", encoding="utf-8") as f:
-            books = json.load(f)
-    except (FileNotFoundError, json.JSONDecodeError):
-        books = []
-
-    # Removes the book with matching ISBN and adds the rest of the books to a new list
-    updated_books = []
-    for book in books:
-        if book.get("isbn") != deleted_book.get("isbn"):
-            updated_books.append()
-    
-    books = updated_books # Update the old list
-
-    # Write to file
-    with open(json_path, "w", encoding="utf-8") as f:
-        json.dump(books, f, ensure_ascii=False, indent=4)
+        # ^ = beginning of string, $ = end of string
+        # options: i = case sensitive matching enabled, "Ape" matches with "APE" & "ape" too
+        return list(collection.find({"author": {"$regex": f"^{query}$", "$options": "i"}}))
